+++ raiden-dapp/src/views/BackupState.vue
@@ -7,6 +7,7 @@
     </v-row>
     <v-list class="backup-state__buttons">
       <v-list-item
+        :disabled="!isConnected"
         class="backup-state__buttons__download-state"
         @click="downloadState = true"
       >
@@ -40,14 +41,17 @@
 
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator';
+import { mapGetters } from 'vuex';
 import DownloadStateDialog from '@/components/DownloadStateDialog.vue';
 
 @Component({
   components: {
     DownloadStateDialog
-  }
+  },
+  computed: { ...mapGetters(['isConnected']) }
 })
 export default class BackupState extends Vue {
+  isConnected!: boolean;
   downloadState: boolean = false;
 }
 </script>